FEAT: Insere estrutura de repetição dos cadastros de protudos, deixa dinamico os status e os itens de info dos produtos
