+++ frontend-gestao-marketplace/src/app/interfaces/product-response.ts
@@ -1,0 +1,9 @@
+export interface IProductResponse {
+  id: number;
+  title: string;
+  price: number;
+  description: string;
+  category: string;
+  status: string;
+  imageBase64: string;
+}
