+++ frontend-gestao-marketplace/src/app/interfaces/products-response.ts
@@ -1,11 +1,6 @@
-export interface IProductResponse {
+import { IProductResponse } from './product-response';
+
+export interface IProductsResponse {
   message: string;
-  data: {
-    id: number;
-    title: string;
-    description: string;
-    category: string;
-    status: string;
-    imageBase64: string;
-  };
+  data: IProductResponse[];
 }
