+++ frontend-gestao-marketplace/src/app/pages/products/products.ts
@@ -1,5 +1,7 @@
-import { Component, inject } from '@angular/core';
+import { Component, inject, OnInit } from '@angular/core';
 import { ProductsService } from '../../services/products';
+import { take } from 'rxjs';
+import { IProductResponse } from '../../interfaces/product-response';
 
 @Component({
   selector: 'app-products',
@@ -7,6 +9,18 @@ import { ProductsService } from '../../services/products';
   templateUrl: './products.html',
   styleUrl: './products.css',
 })
-export class Products {
+export class Products implements OnInit {
+  products: IProductResponse[] = [];
   private readonly _productsService = inject(ProductsService);
+
+  ngOnInit() {
+    this._productsService
+      .getProducts()
+      .pipe(take(1))
+      .subscribe({
+        next: (response) => {
+          this.products = response.data;
+        },
+      });
+  }
 }
